# Writerside/api/api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/admin/subject:
    get:
      tags:
      - admin-subject-controller
      description: Getting subject
      operationId: getSubject
      parameters:
      - name: subjectId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Subject could not be found.
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Found subject
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubjectResponse'
    put:
      tags:
      - admin-subject-controller
      description: Editing subject
      operationId: editSubject
      parameters:
      - name: subjectId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectCreationRequest'
        required: true
      responses:
        "404":
          description: Subject could not be found.
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Edited subject
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Subject area already exists.
          content:
            application/json:
              schema:
                type: string
    post:
      tags:
      - admin-subject-controller
      description: Create new subject
      operationId: createSubject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubjectCreationRequest'
        required: true
      responses:
        "404":
          description: Subject area could not be found.
          content:
            application/json:
              schema:
                type: string
        "201":
          description: Created new subject
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Subject area already exists.
          content:
            application/json:
              schema:
                type: string
    delete:
      tags:
      - admin-subject-controller
      description: Deleting subject
      operationId: deleteSubject
      parameters:
      - name: subjectId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Deleted subject
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubjectResponse'
        "404":
          description: Subject could not be found.
          content:
            application/json:
              schema:
                type: string
...
  /api/admin/admin:
    get:
      tags:
      - admin-admin-controller
      description: Gets Admin
      operationId: getAdmin
      parameters:
      - name: adminId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Admin could not be found.
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Found Admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminResponse'
    put:
      tags:
      - admin-admin-controller
      description: Edit Admin
      operationId: editAdmin
      parameters:
      - name: adminId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSignupRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Admin could not be found.
          content:
            application/json:
              schema:
                type: string
    post:
      tags:
      - admin-admin-controller
      description: Signs up new Admin
      operationId: createAdmin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSignupRequest'
        required: true
      responses:
        "201":
          description: Signed up new Admin
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - admin-admin-controller
      description: Deletes Admin
      operationId: deleteAdmin
      parameters:
      - name: adminId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Admin could not be found.
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Deleted Admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminResponse'
...
  /api/admin/subjects:
    get:
      tags:
      - admin-subject-controller
      description: Getting all subjects
      operationId: getSubjects
      responses:
        "200":
          description: Found all subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubjectResponse'
    delete:
      tags:
      - admin-subject-controller
      description: Deleting list of subjects
      operationId: deleteSubjects
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
                format: int64
        required: true
      responses:
        "200":
          description: Deleted subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubjectResponse'
        "404":
          description: Subject could not be found.
          content:
            application/json:
              schema:
                type: string
...
components:
  schemas:
...
    SubjectResponse:
      type: object
      properties:
        name:
          type: string
...
    SubjectCreationRequest:
      type: object
      properties:
        name:
          type: string
        subjectAreaId:
          type: integer
          format: int64
...
    AdminSignupRequest:
      type: object
      properties:
        firstname:
          type: string
        surname:
          type: string
...
    AdminResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        firstname:
          type: string
        surname:
          type: string
        generatedPassword:
          type: string
        adminId:
          type: integer
          format: int64
...